kernel: fix warnings from stricter compiler flags
Fixes surfaced by the new -Wshadow/-Wconversion/-Wsign-conversion/
-Wcast-align/-Wcast-qual/-Wmissing-prototypes/-Wnull-dereference/
-Wframe-larger-than=512 flag set (added to CFLAGS separately), plus
the null-dereference and frame-size issues LTO's cross-TU view had
been masking after LTO was turned off by default. No logic changes;
type/const-correctness and defensive null-checks only, except two
real bugs found along the way:

- klib/string.c: strncmp's per-byte diff was `signed char` receiving
  an int subtraction of two (default-unsigned on this ABI) chars,
  which can silently misorder results outside +/-127. Widened to
  int, matching strcmp right above it.
- kernel/ipc/sys_msg.c: four ProcessTrackReplyCap() call sites were
  passing `(uint32_t)slot` into a parameter that's actually `Handle`
  (int) -- a stray cast to the wrong type. Removed it; slot is
  already `int`, matching Handle exactly.

Missing prototypes: functions called only from assembly (early(),
exception_dispatch(), exception_exit_pc0_trap(), rdcyc()) get a
forward declaration in the same TU; ValidatePortHandle() and
ZxfCrc32Partial() (file-local only) become static.

cast-qual/cast-align: standard "de-const" libc return patterns
(memchr/strchr/strrchr/bsearch) routed through (uintptr_t); a real
API gap in vm_region_vec_get (no const-taking accessor existed) got
a proper _vec_get_const() added to the shared vector.h macro instead
of casting the const away at the call site.

frame-larger-than=512: boot-only one-shot functions
(FdtEnumerateDevices, dtb_translate_address) had their working
buffers moved to `static` storage instead of the stack; the hot,
concurrent SysWaitAny() syscall path (where `static` would be a real
race) had its four WAITANY_MAX_HANDLES scratch arrays extracted into
a noinline helper so they live in a callee frame instead.

null-dereference: consistent defensive checks added wherever a
*_vec_get()/list_pop_front() accessor's possible NULL return was
being dereferenced unchecked, matching each call site's existing
error-handling/rollback style.

sign-conversion: two root causes accounted for most of it --
arch_reg_set() (see prior commit) for syscall register writes, and
missing `u` suffixes on bitmask-construction constants/macros
(SECTION_SIZE, CACHE_LINE, align_up/align_down, ALIGN_UP_CONST)
where `~(N - 1)` was computed as a negative int before landing in an
unsigned mask. The rest are per-site explicit casts at already-
bounds-checked values (handle table indices, TCB slots, IRQ lines).

+++ include/zuzu/event.h
@@ -8,7 +8,7 @@
 #define KEVENT_MEMMGMT_BIT (1u << 0)
 
 static inline void *ZuzuKEventBind(KEventType eventtype, Handle ntfn, void *args) {
-    return (void *)(VirtAddr)Syscall(SYS_KEVENT_BIND, eventtype, ntfn, (VirtAddr)args, 0);
+    return (void *)(VirtAddr)Syscall(SYS_KEVENT_BIND, eventtype, (uint32_t)ntfn, (VirtAddr)args, 0);
 }
 
 #endif /* EVENT_H */
